+++ Monitor/src/main/java/racecontrol/logging/LoggingPanel.java
@@ -40,7 +40,7 @@ public class LoggingPanel
     }
 
     @Override
-    public void onResize(int w, int h) {
+    public void onResize(float w, float h) {
         table.setPosition(0, 0);
         table.setSize(w, h);
     }
