+++ Monitor/src/main/java/racecontrol/lpgui/gui/LPTextField.java
@@ -385,7 +385,7 @@ public class LPTextField
     }
 
     @Override
-    public void onResize(int w, int h) {
+    public void onResize(float w, float h) {
         updatePresentationText();
     }
